feat: 抽取通用 drill-down 和 drag-move 交互
支持配置面包屑位置 & 整理交互相关的项目目录结构
