+++ src/plots/treemap/interactions/index.ts
@@ -1,2 +1,4 @@
-import './treemap-drill-down';
-import '../../scatter/interaction';
+/** 引入 drag-move 交互 */
+import '../../../interactions/drag-move';
+/** 引入 drill-down 交互 */
+import '../../../interactions/drill-down';
